+++ R/clonalNetwork.R
@@ -30,8 +30,8 @@
 #' are: `gene` (VDJC genes), `nt` (CDR3 nucleotide sequence), `aa` (CDR3 amino 
 #' acid sequence), or `strict` (VDJC + nt). A custom column header can also be used.
 #' @param chain The TCR/BCR chain to use. Use `both` to include both chains 
-#' (e.g., TRA/TRB). Accepted values: `TRA`, `TRB`, `TRG`, `TRD`, `IGH`, `IGL` 
-#' (for both light chains), `both`.
+#' (e.g., TRA/TRB). Accepted values: `TRA`, `TRB`, `TRG`, `TRD`, `IGH`, `IGL`,
+#' `IGK`, `Light` (for both light chains), or `both` (for TRA/B and Heavy/Light).
 #' @param reduction The name of the dimensional reduction of the 
 #' single-cell object.
 #' @param group.by A column header in the metadata or lists to group the analysis